producer/requirements.txt
kafka_python==2.0.2
